rng API: deprecated rng.hpp only. deprecated attributes for old API, deprecated inclusion of rng_device.cuh through rng.cuh
