+++ cpp/include/raft/random/rng.hpp
@@ -19,10 +19,19 @@
  */
 
 /**
- * DISCLAIMER: this file is deprecated (just like rng.cuh) and will produce
- * a compiler warning on include
+ * DISCLAIMER: this file is deprecated: use rng.cuh instead
  */
 
 #pragma once
+
+#if defined(_MSC_VER)
+#pragma message(                                                               \
+  "raft/random/rng.hpp is deprecated and will be removed in a future relase. " \
+  "Please use raft/random/rng_state.hpp or raft/random/rng.cuh instead")
+#else
+#warning \
+  "raft/random/rng.hpp is deprecated and will be removed in a future relase. " \
+  "Please use raft/random/rng_state.hpp or raft/random/rng.cuh instead"
+#endif
 
 #include "rng.cuh"
